+++ app/src/main/java/GraphTest.kt
@@ -1,0 +1,3 @@
+fun main(arr: Array<String>) {
+
+}
