+++ app/src/test/java/GraphTest.kt
@@ -1,0 +1,6 @@
+import junit.framework.TestCase
+import org.junit.Test
+
+class GraphTest {
+
+}
