+++ app/src/test/java/com/example/myapplication/NodeValue.kt
@@ -1,7 +1,7 @@
 package com.example.myapplication
 
 fun getValue(nextNodes: List<Node>): String {
-    val one: String = nextNodes[0].value
-    val two: String = nextNodes[1].value
+    val one: String = nextNodes[0].value.toString()
+    val two: String = nextNodes[1].value.toString()
     return "[$one$two]"
 }
